Add files via upload
Tutorial by BlenderGuru
Learning the basics of Blender, including navigation, modeling, materials, and lighting
Image shows a strawberry donut with sprinkles, placed on a plate, next to an empty mug. Both dishes are designs I have in my own kitchen.
In the future, I may add milk to the mug.